Restructure project into src package and move data
