+++ Other/main_strategy_ugly.py
@@ -2,7 +2,7 @@ import pandas as pd
 import numpy as np
 
 # 1. Read CSV without date parsing
-full_df = pd.read_csv('BATS_QQQ, 60_a45be.csv')
+full_df = pd.read_csv('data/BATS_QQQ, 60_a45be.csv')
 full_df['time'] = pd.to_datetime(full_df['time'].astype(int), unit='s')
 full_df.set_index('time', inplace=True)
 full_df.sort_index(inplace=True)
