+++ data/EURONEXT_DLY_IWDA, 60_6c01f.csv
