+++ data/SIX_DLY_SMI, 60_2d252.csv
