+++ data/XETR_DLY_DAX, 60_dc96e.csv
