+++ data/backtest_results_6_assets.csv
