+++ data/backtest_results_6_assets_input.csv
