+++ src/tradingbot/other/macopy.py
